refactor: Move external secrets provider composable to feature directory
