+++ packages/frontend/editor-ui/src/features/externalSecrets/components/ExternalSecretsProviderCard.ee.vue
@@ -6,7 +6,7 @@ import { useExternalSecretsStore } from '../externalSecrets.ee.store';
 import { useUIStore } from '@/stores/ui.store';
 import { useToast } from '@/composables/useToast';
 import { useI18n } from '@n8n/i18n';
-import { useExternalSecretsProvider } from '@/composables/useExternalSecretsProvider';
+import { useExternalSecretsProvider } from '@/features/externalSecrets/composables/useExternalSecretsProvider';
 import { EXTERNAL_SECRETS_PROVIDER_MODAL_KEY } from '@/constants';
 import { DateTime } from 'luxon';
 import { computed, nextTick, onMounted, toRef } from 'vue';
