+++ packages/frontend/editor-ui/src/features/externalSecrets/composables/useExternalSecretsProvider.ts
@@ -1,10 +1,10 @@
+import type { IUpdateInformation } from '@/Interface';
 import type {
 	ExternalSecretsProvider,
-	IUpdateInformation,
 	ExternalSecretsProviderData,
 	ExternalSecretsProviderProperty,
 	ExternalSecretsProviderState,
-} from '@/Interface';
+} from '../externalSecrets.types';
 import type { ComputedRef, Ref } from 'vue';
 import { computed, ref } from 'vue';
 import { useExternalSecretsStore } from '@/features/externalSecrets/externalSecrets.ee.store';
